refactor: move health, metrics and diagnostics out of the closure
The second group, and the first large one: DiagnosticsSnapshot alone is 246
lines. Closure complexity 548 -> 485, resources.py 3927 -> 3616 lines.

Two things this group forced, both worth recording.

The contract test passed while the extracted classes were broken. It verifies
structure — 34 names, all Resource subclasses, all exposing a verb — and every
one of those held while the code raised NameError at runtime. Thirteen tests in
the full suite caught it. Structure is not behaviour, and the suite is the real
net; the contract test guards the shape of the move, not its correctness.

Diagnostics legitimately reaches across the whole application: scheduler,
storage, oidc, shell_executor. Enumerating each as a context field would be a
fiction, since a diagnostics snapshot is by nature a view over everything. So
ApiContext carries the full manager mapping alongside the named fields, and the
named fields remain the contract for every other group.

The phase-0 lint gate did the work of finding what the move left behind: each
undefined name and orphaned import was reported by name rather than discovered
by running into it.

Refs #669
